refactor(web): Separate plus button menu logic into more component files
