+++ apps/web/src/components/ui-layout/icons-display/AboutMenu.tsx
@@ -1,6 +1,6 @@
 const AboutMenu = () => {
   return (
-    <div className="btn-shadow-dark fixed inset-x-5 top-5 bottom-71 z-50 overflow-auto rounded-lg bg-white shadow-lg">
+    <div className="btn-shadow-dark fixed inset-x-5 top-5 bottom-74 z-50 overflow-auto rounded-lg bg-white shadow-lg">
       <h5 className="px-2 py-1">About CMU Maps</h5>
     </div>
   );
